+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+ - Ignore useless `StreamFactory` parameter in the constructor of `Http\Adapter\Artax\Client` and emit deprecation notice if it's not `null`.
+
 ## v0.1.0
 
 Initial release.